Updates EMC Slurm wrapper
