+++ v6_wrapper/EMC/connection_settings.json
