+++ v6_wrapper/EMC/id_rsa
