+++ v6_wrapper/EMC/known_hosts
