+++ v6_wrapper/EMC/requirements.txt
@@ -1,0 +1,2 @@
+vantage6-client
+vantage6-common
