+++ v6_wrapper/EMC/requirements_master.txt
@@ -1,0 +1,8 @@
+vantage6-client==1.2.3
+vantage6-common==1.2.3
+psycopg2==2.9.3
+protobuf==3.20.*
+tensorflow==2.8.0
+nibabel==3.2.2
+numpy==1.21.5
+scipy==1.7.3
